+++ src/fits_read/error.rs
@@ -6,6 +6,8 @@
 Errors associated with reading in fits files.
  */
 
+use std::path::PathBuf;
+
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -14,7 +16,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\nCouldn't open {fits_filename}: {fits_error}")]
     Open {
         fits_error: fitsio::errors::Error,
-        fits_filename: String,
+        fits_filename: PathBuf,
         source_file: &'static str,
         source_line: u32,
     },
@@ -23,7 +25,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Couldn't find key {key}")]
     MissingKey {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -32,7 +34,7 @@ pub enum FitsError {
     /// Error describing a HDU that couldn't be used as an image (e.g. `HduInfo::ImageInfo`).
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Tried to use as an image, but not an image")]
     NotImage {
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -42,7 +44,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Couldn't read a long string from {key}")]
     LongString {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -52,7 +54,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: {fits_error}")]
     Fitsio {
         fits_error: fitsio::errors::Error,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -62,7 +64,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\nCouldn't parse {key} in {fits_filename} HDU {hdu_num}")]
     Parse {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,