+++ src/gpubox_files/mod.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Functions for organising and checking the consistency of gpubox files.
-*/
+//! Functions for organising and checking the consistency of gpubox files.
+
 pub mod error;
 
 use std::collections::BTreeMap;