+++ src/metafits_context/error.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Errors associated with reading in metafits files.
-*/
+//! Errors associated with reading in metafits files.
+
 use crate::MWAMode;
 use thiserror::Error;