+++ src/timestep/test.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Unit tests for timestep metadata
-*/
+//! Unit tests for timestep metadata
+
 #[cfg(test)]
 use super::*;
 use crate::gpubox_files::GpuboxTimeMap;