+++ src/voltage_context/mod.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-The main interface to MWA voltage data.
- */
+//! The main interface to MWA voltage data.
+
 use crate::coarse_channel::*;
 use crate::error::*;
 use crate::metafits_context::*;